First, fixed a bug when calculating soft point correspondences where the search box size wasn't getting set properly
Second, target find test prg is now using the correspondence to create hard matches. I need some better logic here, though
as it runs into problem if multiple contours are found for the same object. At a minimum, I need it to pick the one that has lived the longest.